+++ libsakura/src/libsakura/optimized_implementation_factory_impl.h
@@ -39,16 +39,16 @@ public:
 	}
 	virtual void SubtractBaselinePolynomial(size_t num_data,
 			float const data[/*num_data*/],
-			bool const mask[/*num_data*/], size_t order,
-			float clipping_threshold_sigma, size_t num_fitting_max,
+			bool const mask[/*num_data*/], uint8_t order,
+			float clipping_threshold_sigma, uint8_t num_fitting_max,
 			bool get_residual, float out[/*num_data*/]) const;
-	virtual void GetBaselineModel(size_t num_each_basis, size_t order,
+	virtual void GetBaselineModel(size_t num_each_basis, uint8_t order,
 			double model[/*(order+1)*num_each_basis*/]) const;
 	virtual void DoSubtractBaseline(size_t num_data,
 			float const data[/*num_data*/],
 			bool const mask[/*num_data*/], size_t num_model_bases,
 			double const model[/*num_model_bases*num_data*/],
-			float clipping_threshold_sigma, size_t num_fitting_max,
+			float clipping_threshold_sigma, uint8_t num_fitting_max,
 			bool get_residual, float out[/*num_data*/]) const;
 };
 
@@ -58,16 +58,16 @@ public:
 	}
 	virtual void SubtractBaselinePolynomial(size_t num_data,
 			float const data[/*num_data*/],
-			bool const mask[/*num_data*/], size_t order,
-			float clipping_threshold_sigma, size_t num_fitting_max,
+			bool const mask[/*num_data*/], uint8_t order,
+			float clipping_threshold_sigma, uint8_t num_fitting_max,
 			bool get_residual, float out[/*num_data*/]) const;
-	virtual void GetBaselineModel(size_t num_each_basis, size_t order,
+	virtual void GetBaselineModel(size_t num_each_basis, uint8_t order,
 			double model[/*(order+1)*num_each_basis*/]) const;
 	virtual void DoSubtractBaseline(size_t num_data,
 			float const data[/*num_data*/],
 			bool const mask[/*num_data*/], size_t num_model_bases,
 			double const model[/*num_model_bases*num_data*/],
-			float clipping_threshold_sigma, size_t num_fitting_max,
+			float clipping_threshold_sigma, uint8_t num_fitting_max,
 			bool get_residual, float out[/*num_data*/]) const;
 };
